league data extraction
This file works into the Riot API to extract the data I need for my project and puts it into a Pandas dataframe. I am still trying to work on getting them in the finished form but this is what I have so far. In this file, I also utilized Community Dragon to properly view some of the items, summoner spells, and runes in their appropriate names so they don't just look like numbers.